Доработка в стиле MVP
