+++ js/game/game-view.js
@@ -1,22 +1,25 @@
 
-import AbstractView from './view';
-import HeaderView from './header-view';
-import runGame from './game';
+import AbstractView from '../view';
+// import Application from '../application';
+// import HeaderView from './header-view';
 
-export default class GameView extends AbstractView {
+class GameView extends AbstractView {
 
   constructor() {
     super();
-    this.header = new HeaderView();
+//    this.header = new HeaderView(gameState);
   }
 
   getMarkup() {
-    return `${this.header.getMarkup(true)}
-      <div class="game"></div>`;
+//    return `${this.header.getMarkup(true)}
+//      <div class="game"></div>`;
+    return '<div class="game"></div>';
   }
 
   bindHandlers() {
-    this.element.onshow = () => runGame();
+//    this.element.querySelector('.header__back').onclick = (evt) => Application.showGreeting();
   }
 
 }
+
+export default () => new GameView().element;
