+++ js/game/level-view.js
@@ -1,5 +1,5 @@
 
-import AbstractView from './view';
+import AbstractView from '../view';
 
 export default class LevelView extends AbstractView {
