+++ js/screen/stats-screen.js
@@ -1,14 +1,15 @@
 
 import AbstractView from '../view';
-// import Application from '../application';
+import Application from '../application';
 import HeaderView from '../game/header-view';
-import {getStats} from '../data/markup';
-import {gameState} from '../data/data';
+import {getStats} from '../game/game';
+import {initialState} from '../game/game';
 
 class StatsView extends AbstractView {
 
-  constructor() {
+  constructor(gameState) {
     super();
+    this._state = gameState;
     this.header = new HeaderView();
   }
 
@@ -19,11 +20,11 @@ class StatsView extends AbstractView {
       lives: 50
     };
 
-    const score = gameState.answers.correct * factor.points;
-    const scoreString = (score > 0) ? gameState.answers.correct + '&nbsp;×&nbsp;' + factor.points : '';
-    const fastBonus = gameState.answers.fast * factor.speed;
-    const livesBonus = gameState.lives.left * factor.lives;
-    const slowPenalty = gameState.answers.slow * factor.speed;
+    const score = this._state.answers.correct * factor.points;
+    const scoreString = (score > 0) ? this._state.answers.correct + '&nbsp;×&nbsp;' + factor.points : '';
+    const fastBonus = this._state.answers.fast * factor.speed;
+    const livesBonus = this._state.lives.left * factor.lives;
+    const slowPenalty = this._state.answers.slow * factor.speed;
 
     const scoreTotal = (score > 0) ? score + fastBonus + livesBonus - slowPenalty : 0;
 
@@ -35,7 +36,7 @@ class StatsView extends AbstractView {
         <table class="result__table">
         <tr>
           <td class="result__number">1.</td>
-          <td colspan="2">${getStats(gameState.levels)}</td>
+          <td colspan="2">${getStats(this._state.levels)}</td>
           <td class="result__points">${scoreString}</td>
           <td class="result__total">${score}</td>
         </tr>`;
@@ -44,7 +45,7 @@ class StatsView extends AbstractView {
         result += `<tr>
             <td></td>
             <td class="result__extra">Бонус за скорость:</td>
-            <td class="result__extra">${gameState.answers.fast}&nbsp;<span class="stats__result  stats__result--fast"></span></td>
+            <td class="result__extra">${this._state.answers.fast}&nbsp;<span class="stats__result  stats__result--fast"></span></td>
             <td class="result__points">×&nbsp;${factor.speed}</td>
             <td class="result__total">${fastBonus}</td>
           </tr>`;
@@ -53,7 +54,7 @@ class StatsView extends AbstractView {
         result += `<tr>
             <td></td>
             <td class="result__extra">Бонус за жизни:</td>
-            <td class="result__extra">${gameState.lives.left}&nbsp;<span class="stats__result  stats__result--heart"></span></td>
+            <td class="result__extra">${this._state.lives.left}&nbsp;<span class="stats__result  stats__result--heart"></span></td>
             <td class="result__points">×&nbsp;${factor.lives}</td>
             <td class="result__total">${livesBonus}</td>
           </tr>`;
@@ -62,7 +63,7 @@ class StatsView extends AbstractView {
         result += `<tr>
             <td></td>
             <td class="result__extra">Штраф за медлительность:</td>
-            <td class="result__extra">${gameState.answers.slow}&nbsp;<span class="stats__result  stats__result--slow"></span></td>
+            <td class="result__extra">${this._state.answers.slow}&nbsp;<span class="stats__result  stats__result--slow"></span></td>
             <td class="result__points">×&nbsp;${factor.speed}</td>
             <td class="result__total">-${slowPenalty}</td>
           </tr>`;
@@ -78,9 +79,9 @@ class StatsView extends AbstractView {
   }
 
   bindHandlers() {
-//    document.querySelector('.header__back').onclick = () => Application.showGreeting();
+    this.element.querySelector('.header__back').onclick = (evt) => Application.showGreeting();
   }
 
 }
 
-export default () => new StatsView().element;
+export default () => new StatsView(initialState).element;
